Create Fun With Strings
A program that will perform a few different String operations. The program prompts for a name and then tells the user how many letters are in the name, what the first letter of the name is, and what the last letter of the name is. Then it ends with a goodbye message.